feat(adr-061): open extract package, move castxml name helpers (Phase 5 item 1)
Starts ADR-061 Phase 5 item 1 ("split CastXML and Clang parsing by entity
and shared parser context"). Opens abicheck/extract/ with a scoped
AGENTS.md and its first tenant, extract/headers/castxml/names.py: the
vtable-index, mangled-name, and synthetic-key helpers that sat as
module-level functions above _CastxmlParser in dumper_castxml.py. Each is
a pure computation over a string or single XML element with no dependency
on the parser's shared id-map state, which is why this piece could move
before the shared parser context the rest of the split needs.

dumper_castxml.py imports and re-exports every moved name unchanged, so
every existing caller (dumper.py, diff_symbols.py, diff_templates.py,
tu_merge.py, dumper_hybrid.py, finding_identity_ctor_dtor.py,
buildsource/ctor_export_match.py, and tests importing from
abicheck.dumper/abicheck.dumper_castxml directly) keeps resolving. File
drops from 1934 to 1822 lines.

Verified: check_architecture.py 0 errors, mypy clean, ruff clean, and
every test file that imports the moved names passes unchanged.

Updates the ADR's status line and item-1 entry to record this honestly as
started, not done — the stateful entity-by-entity split of
_CastxmlParser/_ClangAstParser is unstarted follow-up work.
