+++ abicheck/extract/__init__.py
@@ -1,0 +1,23 @@
+# Copyright 2026 Nikolay Petrov
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""ADR-061 ``extract`` responsibility package.
+
+Reads binary, debug, header, build, and source evidence into facts owned by
+``abicheck.model``. Small and inert per ADR-061 D6 — no plugin registration,
+no environment inspection, no product logic. See ``AGENTS.md`` in this
+directory for the package contract.
+"""
+
+from __future__ import annotations
