+++ abicheck/extract/headers/__init__.py
@@ -1,0 +1,22 @@
+# Copyright 2026 Nikolay Petrov
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Header-AST parser backends (castxml, clang) under ``extract``.
+
+Each backend is its own sub-package, split by parsed entity and shared
+parser context per ADR-061 D9. This ``__init__.py`` stays inert — it holds
+no shared logic of its own.
+"""
+
+from __future__ import annotations
